FindByIdAsync now parses to long instead of int

+++ Auth/Storage/Strasnote.Auth.Data/UserStore.cs
@@ -52,7 +52,7 @@ namespace Strasnote.Auth.Data
 		{
 			ThrowIfDisposed();
 
-			return userContext.RetrieveByIdAsync<UserEntity>(int.Parse(userId));
+			return userContext.RetrieveByIdAsync<UserEntity>(long.Parse(userId));
 		}
 
 		/// <inheritdoc/>
